+++ lib/core/widgets/custom_text_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:nova_store_app/core/theming/app_text_styles.dart';
+import 'package:nova_store_app/core/theme/app_text_styles.dart';
 import 'package:nova_store_app/core/widgets/custom_button.dart';
 
 class CustomTextButton extends StatelessWidget {
