+++ lib/features/auth/presentation/views/widgets/dont_have_an_account.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:nova_store_app/core/routing/routes_paths.dart';
-import 'package:nova_store_app/core/theming/app_colors.dart';
-import 'package:nova_store_app/core/theming/app_text_styles.dart';
+import 'package:nova_store_app/core/theme/app_colors.dart';
+import 'package:nova_store_app/core/theme/app_text_styles.dart';
 import 'package:nova_store_app/core/widgets/clickable_text.dart';
 
 class DontHaveAnyAccount extends StatelessWidget {
