+++ lib/features/auth/presentation/views/widgets/email_text_form_field.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:nova_store_app/core/theming/app_colors.dart';
+import 'package:nova_store_app/core/theme/app_colors.dart';
 import 'package:nova_store_app/core/utils/assets.dart';
 import 'package:nova_store_app/core/utils/form_validators.dart';
 import 'package:nova_store_app/core/widgets/custom_text_form_field.dart';
