+++ lib/features/auth/presentation/views/widgets/enter_email_section.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:nova_store_app/core/theming/app_colors.dart';
-import 'package:nova_store_app/core/theming/app_text_styles.dart';
+import 'package:nova_store_app/core/theme/app_colors.dart';
+import 'package:nova_store_app/core/theme/app_text_styles.dart';
 import 'package:nova_store_app/features/auth/presentation/views/widgets/email_text_form_field.dart';
 
 class EnterEmailSection extends StatelessWidget {
