+++ lib/features/auth/presentation/views/widgets/login_form.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:nova_store_app/core/theming/app_colors.dart';
+import 'package:nova_store_app/core/theme/app_colors.dart';
 import 'package:nova_store_app/core/utils/spacing.dart';
 import 'package:nova_store_app/features/auth/presentation/views/widgets/auth_app_bar.dart';
 import 'package:nova_store_app/features/auth/presentation/views/widgets/dont_have_an_account.dart';
