+++ lib/features/onboarding/presentation/views/widgets/onboarding_bottom_sheet/onboading_bottom_sheet.dart
@@ -4,8 +4,8 @@ import 'package:flutter_svg/svg.dart';
 import 'package:nova_store_app/core/di/service_locator.dart';
 import 'package:nova_store_app/core/extensions/clamping.dart';
 import 'package:nova_store_app/core/helpers/on_boarding_cache_helper.dart';
-import 'package:nova_store_app/core/theming/app_colors.dart';
-import 'package:nova_store_app/core/theming/app_text_styles.dart';
+import 'package:nova_store_app/core/theme/app_colors.dart';
+import 'package:nova_store_app/core/theme/app_text_styles.dart';
 import 'package:nova_store_app/core/utils/assets.dart';
 import 'package:nova_store_app/core/utils/spacing.dart';
 import 'package:nova_store_app/features/onboarding/presentation/views/widgets/onboarding_bottom_sheet/get_started_button.dart';
