+++ lib/features/onboarding/presentation/views/widgets/yellow_badge.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:nova_store_app/core/theming/app_colors.dart';
+import 'package:nova_store_app/core/theme/app_colors.dart';
 
 class YellowBadge extends StatelessWidget {
   const YellowBadge({
